\name{arrayMissPattern}
\alias{arrayMissPattern}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Software for Missing Pattern Analysis of Microarray data}
\description{
arrayMissPattern is a software to explore missing patterns in microarray data
This software provides its own GUI to help biologists who are not familiar with R commands.
}
\usage{
arrayMissPattern()
}


\author{ Eun-kyung Lee, Dankyu Yoon, Taesung Park}
\examples{
arrayMissPattern()
}
\keyword{ dplot }% at least one, from doc/KEYWORDS
\keyword{ utilities }% __ONLY ONE__ keyword per line
